Day 2: Setup auth using Laravel Sanctum. Created user routes, model and controller. Setup middlewares to block certain routes from unauthenticated users.  Created a relationship between the user and the products where:
- Each product must be created by an authenticated user.
- A product must have only one user linked to it, but a user can have many products.

Created a policy that ensures only the user that owns a post can update or delete it.

+++ app/Http/Controllers/ProductController.php
@@ -4,9 +4,18 @@ namespace App\Http\Controllers;
 
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Routing\Controllers\HasMiddleware;
 
-class ProductController extends Controller
+class ProductController extends Controller implements HasMiddleware
 {
+    public static function middleware()
+    {
+        return[
+            new Middleware('auth:sanctum', except: ['index', 'show'])
+        ];
+    }
     /**
      * Display a listing of the resource.
      */
@@ -28,7 +37,7 @@ class ProductController extends Controller
             'amount_sold' => 'required'
         ]);
 
-        $product = Product::create($fields);
+        $product = $request->user()->products()->create($fields);
 
         return $product;
     }
@@ -46,6 +55,8 @@ class ProductController extends Controller
      */
     public function update(Request $request, Product $product)
     {
+        Gate::authorize('modify', $product);
+
         $fields = $request->validate([
             'name' => 'required|max:20',
             'description' => 'required|max:255',
@@ -64,6 +75,8 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
+        Gate::authorize('modify', $product);
+        
         $product->delete();
 
         return ['message' => "The product was deleted"];
